[Jones] add publishing folder
