+++ sql_refactor/README.md
@@ -5,7 +5,7 @@ This library contains a class, `Refactor`, for refactoring a SQL with a given kn
 The knowledge is a dictionary containing the mapping of `old_table`-`new_table` and corresponding `old_column`-`new_column`.
 
 The structure of the knowledge:
-```
+```json
 {
     "old_table_name":
     {
@@ -22,4 +22,28 @@ The structure of the knowledge:
     .
     .
 }
+```
+
+## How to use
+
+Code sample:
+```py
+from sql_refactor import Refactor
+
+# refer to Structure of Knowledge
+knowledge = KNOWLEDGE
+
+refactor = Refactor(knowledge)
+refactor.refactor('''SELECT column_1, column_2 FROM old_table_name''')
+
+new_query = refactor.result()
+```
+
+Result:
+```sql
+SELECT
+  new_column_1 AS column_1,
+  new_column_2 AS column_2
+FROM
+  `new_table_name`
 ```
